Add authenticated project creat/delete/get

+++ src/main/java/gatech/edu/ppmtool/services/ProjectService.java
@@ -2,12 +2,14 @@ package gatech.edu.ppmtool.services;
 
 import gatech.edu.ppmtool.domain.Backlog;
 import gatech.edu.ppmtool.domain.Project;
+import gatech.edu.ppmtool.domain.User;
 import gatech.edu.ppmtool.exceptions.ProjectIdException;
 import gatech.edu.ppmtool.repository.BacklogRepository;
 import gatech.edu.ppmtool.repository.ProjectRepository;
-import jdk.nashorn.internal.runtime.arrays.IteratorAction;
 import org.springframework.stereotype.Service;
 import org.springframework.beans.factory.annotation.Autowired;
+
+import java.security.Principal;
 
 @Service
 public class ProjectService {
@@ -17,7 +19,7 @@ public class ProjectService {
     @Autowired
     private BacklogRepository backlogRepository;
 
-    public Project saveOrUpdateProject(Project project) {
+    public Project saveOrUpdateProject(Project project, String username) {
         try {
             project.setProjectId(project.getProjectId().toUpperCase());
             if (project.getId() == null) {
@@ -25,6 +27,7 @@ public class ProjectService {
                 backlog.setProject(project);
                 backlog.setProjectId(project.getProjectId());
                 project.setBacklog(backlog);
+                project.setManagerUsername(username);
             } else {
                 project.setBacklog(backlogRepository.findByProjectId(project.getProjectId()));
             }
@@ -34,22 +37,29 @@ public class ProjectService {
         }
     }
 
-    public Project readProjectByProjectId(String projectId) {
+    public Project readProjectByProjectId(String projectId, Principal principal) {
         Project res = projectRepository.findByProjectId(projectId);
         if (res == null) {
             throw new ProjectIdException("Project id '" + projectId + "' doesn't exist");
         }
+        System.out.println(principal.getName());
+        if (!res.getManagerUsername().equals(principal.getName())) {
+            throw new ProjectIdException("User '" + principal.getName() + "' doesn't manage project '" + res.getProjectId() + "'");
+        }
         return res;
     }
 
-    public Iterable<Project> readAllProjects() {
-        return projectRepository.findAll();
+    public Iterable<Project> readAllProjects(Principal principal) {
+        return projectRepository.findAllByManagerUsername(principal.getName());
     }
 
-    public void deleteProjectByProjectId(String projectId) {
+    public void deleteProjectByProjectId(String projectId, Principal principal) {
         Project proj = projectRepository.findByProjectId(projectId.toUpperCase());
         if (proj == null) {
             throw new ProjectIdException("Cannot delete project '" + projectId + "' (doesn't exist)");
+        }
+        if (!proj.getManagerUsername().equals(principal.getName())) {
+            throw new ProjectIdException("User '" + principal.getName() + "' doesn't manage project '" + proj.getProjectId() + "'");
         }
         projectRepository.delete(proj);
     }
